backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python==4.8.1.78
tensorflow==2.14.0
Pillow==10.1.0
python-jose==3.3.0
websockets==12.0
torch==2.0.1
torchvision==0.15.2
